Fixed 16 more warnings :D (don't know how to fix remaining warnings)
With Rik Ruiten

+++ source/Game/Stations/PowerStation.cpp
@@ -281,12 +281,12 @@ void PowerStation::update()
 	int weapon		= context.GetPower(ST_WEAPON);
 	int navigation	= context.GetPower(ST_NAVIGATION);
 
-	context.powerPoolText->setText((varToString("Power Pool:\n", context.powerPool, "%")).c_str());
+	context.powerPoolText->setText((varToString("Power Pool:\n", (float)context.powerPool, "%")).c_str());
 
-	context.helmStatus->setText((varToString(		"Helm power status: ", helm, "%")).c_str());
-	context.defenceStatus->setText((varToString(	"Defence power status: ", defence, "%")).c_str());
-	context.weaponStatus->setText((varToString(		"Weapon power status: ", weapon, "%")).c_str());
-	context.navigationStatus->setText((varToString(	"Navigation power status: ", navigation, "%")).c_str());
+	context.helmStatus->setText((varToString(		"Helm power status: ", (float)helm, "%")).c_str());
+	context.defenceStatus->setText((varToString(	"Defence power status: ", (float)defence, "%")).c_str());
+	context.weaponStatus->setText((varToString(		"Weapon power status: ", (float)weapon, "%")).c_str());
+	context.navigationStatus->setText((varToString(	"Navigation power status: ", (float)navigation, "%")).c_str());
 
 	//Checks the power percentage and assigns the text a color indicating the amount of power available to that station.
 	changeColorAccordingToPowerStatus(*context.helmStatus, (float)helm);
